fix(publish): 修复复制文件时的目标文件已存在问题是链接文件的情况
- 在复制文件之前，检查目标文件是否存在
- 如果存在，则删除目标文件，以确保复制操作成功

+++ src/publish-package.ts
@@ -207,7 +207,7 @@ function _4copyFile(pattern: RegExp, fileName: string, meta: InternalPublishMeta
   const destFile = path.join(meta.pkgRoot, fileName);
 
   // 如果目标文件是链接文件，如果不删除则复制不成功
-  fs.unlinkSync(destFile);
+  if (fs.existsSync(destFile)) fs.unlinkSync(destFile);
 
   // 复制文件
   fs.copyFileSync(sourceFile, destFile);
